refactor(fl): batch-relocate time_alpha, clockless_controller_impl, str_ui
Three independent file moves into existing module subfolders, batched into
one commit since they're trivial and share the same mechanical sweep:

  - fl/time_alpha.{h,cpp.hpp}        -> fl/math/   (alongside math.h, ease.h)
  - fl/clockless_controller_impl.h   -> fl/chipsets/   (alongside other clockless types)
  - fl/str_ui.cpp.hpp                -> fl/ui/string.cpp.hpp   (UI-side bridge into fl::string)

Plus the matching test relocation per TestPathStructureChecker:
  - tests/fl/time_alpha.hpp -> tests/fl/math/time_alpha.hpp

Updates unity-build aggregators (src/fl/_build.cpp.hpp, src/fl/build/fl.cpp),
adds entries to the destination module's _build.cpp.hpp, and sweeps all
includes across src/, tests/, examples/, and README.md.
