+++ README.md
@@ -419,7 +419,7 @@ Lots of improvements in this release, read the full [change list here](https://g
     * [Wave Simulation Library](https://github.com/FastLED/FastLED/blob/master/src/fl/wave_simulation.h)
   * [FireCylinder](https://github.com/FastLED/FastLED/blob/master/examples/FireCylinder/FireCylinder.ino)
     * Wraps around so that (0,y) ~= (width-1,y)
-  * [TimeAlpha](https://github.com/FastLED/FastLED/blob/master/src/fl/time_alpha.h)
+  * [TimeAlpha](https://github.com/FastLED/FastLED/blob/master/src/fl/math/time_alpha.h)
     * Precision control of animations with time-based alpha transition.
