+++ examples/Audio/simple/fx_audio.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "fl/time_alpha.h"
+#include "fl/math/time_alpha.h"
 #include "fl/math/math.h"
 
 /// Tracks a smoothed peak with attack, decay, and output-inertia time-constants.
