+++ examples/XYPath/direct.h
@@ -19,7 +19,7 @@ all the UI elements you see below.
 #include "fl/gfx/draw_visitor.h"
 #include "fl/math/math.h"
 #include "fl/gfx/raster.h"
-#include "fl/time_alpha.h"
+#include "fl/math/time_alpha.h"
 #include "fl/ui.h"
 #include "fl/gfx/xypath.h"
 #include "fl/fx/time.h"
