+++ src/fl/_build.cpp.hpp
@@ -10,8 +10,6 @@
 #include "fl/fastled_internal.cpp.hpp"
 #include "fl/fltest.cpp.hpp"
 #include "fl/static_constexpr_defs.cpp.hpp"
-#include "fl/str_ui.cpp.hpp"
-#include "fl/time_alpha.cpp.hpp"
 
 // begin sub directory includes
 #include "fl/audio/_build.cpp.hpp"
