+++ src/platforms/clockless.h
@@ -50,4 +50,4 @@
 
 // Template alias to ClocklessController (platform-specific or generic blocking)
 // This must come AFTER all clockless drivers are included
-#include "fl/clockless_controller_impl.h"
+#include "fl/chipsets/clockless_controller_impl.h"
