+++ tests/fl/math/time_alpha.hpp
@@ -1,7 +1,7 @@
 // test_transition_ramp.cpp
 // g++ --std=c++11 test_transition_ramp.cpp
 
-#include "fl/time_alpha.h"
+#include "fl/math/time_alpha.h"
 #include "fl/stl/stdint.h"
 #include "test.h"
